pyproject.toml
[tool.poetry]
name = "drf-lookup-fields"
version = "1.0.0"
description = "A package supplying tools for custom foreign-key fields lookup in drf-created WebAPI."
authors = ["IIIT <github@iiit.pl>"]
readme = "README.md"
license = "BSD-2-Clause"
repository = "https://github.com/innovationinit/drf-lookup-fields"
packages = [
    { include = "lookup_fields" },
]
include = [
    "README.md",
    "LICENSE",
    "lookup_fields/locale/**",
]

[tool.poetry.dependencies]
python = ">=3.7,<3.10"
Django = ">=2.0,<3.0"
djangorestframework = ">=3.3.2"

[tool.poetry.dev-dependencies]
tox = "^3.24.5"

[tool.coverage.run]
branch = true
source = ["lookup_fields"]

[tool.coverage.report]
omit = ["*/testsproject/*"]
exclude_lines = ["pragma: no cover"]

[tool.coverage.html]
directory = "coverage"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
